fix: use final extension structure
Initial work was done when extensions were brand new, but things have changed
since. This consolidates the former JSON files into a single TOML file, and
configures prettier to maybe work. (It doesn't, but it tries to!)

PHP injections are still broken, though. This tries to use the newish php_only
tree-sitter parser, but to no avail. Also, there seems to be an issue with the
Blade parser and how it handles directive parameters: they're not treated as a
single node, but as a bunch of sibling nodes.